fix: after resolve merge conflict

+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmStorage.java
@@ -19,7 +19,7 @@ public interface FilmStorage {
 
     List<Film> findByIds(List<Long> ids);
 
-    List<Film> findFilmsByDirector(int directorId, String sortBy);
+    List<Film> findFilmsByDirector(long directorId, String sortBy);
 
     List<Film> findByNameContainingIgnoreCase(String query);
